can't randomize still
