+++ src/Pages/trivia.JSON
@@ -1,0 +1,1 @@
+{"response_code":0,"results":[{"category":"History","type":"multiple","difficulty":"hard","question":"In the year 1900, what were the most popular first names given to boy and girl babies born in the United States?","correct_answer":"John and Mary","incorrect_answers":["Joseph and Catherine","William and Elizabeth","George and Anne"]},{"category":"Entertainment: Video Games","type":"multiple","difficulty":"hard","question":"Which one of the following titles by Valve is not based on a Community Mod?","correct_answer":"Ricochet","incorrect_answers":["Day of Defeat","Counter-Strike","Alien Swarm"]},{"category":"Entertainment: Music","type":"multiple","difficulty":"hard","question":"Which of the following is NOT a real song from the band Thousand Foot Krutch?","correct_answer":"Limitless Fury","incorrect_answers":["Let The Sparks Fly","Down","Give Up The Ghost"]},{"category":"Entertainment: Video Games","type":"multiple","difficulty":"hard","question":"In what year was the game &quot;Eliminator&quot; released?","correct_answer":"1998","incorrect_answers":["2000","1999","2001"]},{"category":"Science & Nature","type":"multiple","difficulty":"hard","question":"What element on the periodic table has 92 electrons?","correct_answer":"Uranium","incorrect_answers":["Sulfur","Hydrogen","Iron"]},{"category":"Entertainment: Music","type":"multiple","difficulty":"hard","question":"Who wrote the song &quot;You Know You Like It&quot;?","correct_answer":"AlunaGeorge","incorrect_answers":["DJ Snake","Steve Aoki","Major Lazer"]},{"category":"Science: Gadgets","type":"multiple","difficulty":"hard","question":"Who built the first laser?","correct_answer":"Theodore Harold Maiman","incorrect_answers":["Nikola Tesla","Jack Kilby","Edith Clarke"]},{"category":"Science & Nature","type":"multiple","difficulty":"hard","question":"Muscle fiber is constructed of bundles small long organelles called what?","correct_answer":"Myofibrils","incorrect_answers":["Epimysium","Myofiaments","Myocardium"]},{"category":"Entertainment: Video Games","type":"multiple","difficulty":"hard","question":"What is the 4th boss in the 1997 video game &quot;Crash Bandicoot 2: Cortex Strikes Back&quot;?","correct_answer":"Dr. N. Gin","incorrect_answers":["Dr. Neo Cortex","Komodo Brothers","Tiny Tiger"]},{"category":"Entertainment: Film","type":"multiple","difficulty":"hard","question":"Which of the following films was directed by Ivan Reitman, written by Gary Ross, featured Kevin Kline, and was released in 1993?","correct_answer":"Dave","incorrect_answers":["John","Will","Carl"]}]}
